requirements.txt
anthropic>=0.39.0
openai>=1.58.0
opencv-python>=4.10.0
numpy>=1.26.0
sounddevice>=0.5.0
scipy>=1.14.0
keyboard>=0.13.5
PyQt6>=6.7.0
python-dotenv>=1.0.0
Pillow>=10.4.0
tavily-python>=0.5.0
webrtcvad>=2.0.10
faster-whisper>=1.0.0
pytest>=8.0.0
